+++ src/renderer/index.js
@@ -1,7 +1,9 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import 'antd/dist/antd.css';
-import './now-note.css';
+
+import 'antd/es/style/themes/default.less';
+import './now-note.less';
 import {App} from './components/App.jsx';