+++ webpack.renderer.config.js
@@ -8,6 +8,22 @@ rules.push({
   use: [{ loader: 'style-loader' }, { loader: 'css-loader' }],
 });
 
+rules.push({
+  test: /\.less$/i,
+  use: [
+    { loader: 'style-loader' }, 
+    { loader: 'css-loader' },
+    { 
+      loader: 'less-loader',
+      options: {
+        lessOptions: {
+            javascriptEnabled: true,
+        }
+      } 
+    },
+  ]
+});
+
 module.exports = {
   // Put your normal webpack config below here
   module: {